refactor: plugin models
