+++ smarter/smarter/apps/plugin/models/exceptions.py
@@ -1,0 +1,10 @@
+"""
+Exceptions for the Plugin app.
+"""
+
+# python stuff
+from smarter.common.exceptions import SmarterValueError
+
+
+class PluginDataValueError(SmarterValueError):
+    """Custom exception for PluginData SQL errors."""
